+++ buildapp/build_app_windows_x86.sh
@@ -1,8 +1,15 @@
 #!/bin/bash
+if (( $EUID == 0 )); then echo "The build process is not meant to be run with root privilege, exiting now." >&2; exit 1; fi
+
 cd "${0%/*}"
 SRCFILES="terrarumwindows_x86"
 DESTDIR="TerrarumWindows.x86"
 RUNTIME="runtime-windows-x86"
+
+if [ ! -d "../assets_release" ]; then
+    echo "'assets_release' does not exist; prepare the assets for the release and put them into the assets_release directory, exiting now." >&2
+    exit 1
+fi
 
 # Cleanup
 rm -rf $DESTDIR || true
@@ -15,7 +22,8 @@ cp $SRCFILES/Terrarum.bat $DESTDIR/
 cp -r "../out/$RUNTIME" $DESTDIR/
 
 # Copy over all the assets and a jarfile
-cp -r "../assets" $DESTDIR/
+cp -r "../assets_release" $DESTDIR/
+mv $DESTDIR/assets_release $DESTDIR/assets
 cp -r "../out/TerrarumBuild.jar" $DESTDIR/assets/
 
 # Temporary solution: zip everything
